+++ src/pages/admin/Produtos/index.js
@@ -1,14 +1,34 @@
-import React from 'react';
+import React, {useState,useEffect} from 'react';
 import { BiFace,BiHomeAlt,BiLayerPlus, BiPackage,BiBullseye } from "react-icons/bi";
 import { IoIosAddCircle,IoIosExit} from "react-icons/io";
+import {FaTrash} from 'react-icons/fa';
 import {Link} from 'react-router-dom';
 import './style.css'
 import Logo from '../../../assets/logored.svg';
 import Graph from '../../../assets/Dashboard/graph.svg';
-import data from '../Dashboard/api';
+import api from '../../../service/api';
 
 export default function AddProdutos() {
 
+
+  const [ produtos, setProdutos ] = useState([])
+
+  async function DeletarProduto(prop){
+    await api.get(`/excluir/${prop}`)
+  }
+
+  useEffect(()=>{
+    async function getProdutos(){
+      const {data} = await api.get('/listProduto')
+      return setProdutos(data.produtos);
+    }
+    getProdutos();
+  },[produtos])
+
+
+  
+
+  //Para o componente de metas (por enquanto estático);
   const meta = false;
 
   return (
@@ -30,7 +50,7 @@ export default function AddProdutos() {
 
 <div className="funcoes">
       <Link to="/dashboard" className="itens"> <BiHomeAlt size="30px"/> </Link>
-      <Link to="/addprodutos" className="itens"> <BiPackage size="30px"/> </Link>
+      <Link to="/produtos" className="itens"> <BiPackage size="30px"/> </Link>
       <Link to="/clientes" className="itens"> <BiFace size="30px"/> </Link>
 </div>
 
@@ -39,17 +59,20 @@ export default function AddProdutos() {
       <h1 style={{color: '#820E0E', fontSize: 20, marginBottom: 30}}>Produtos</h1>
 
     <div className="produtosCadastrados">
-      <Link className="add">
+      <Link className="add" to="/addprodutos">
         <IoIosAddCircle size="73px" />
         <h1>Adicionar Produto</h1>
       </Link>
-      { data.map(e=>{
+
+      { produtos.map(e=>{
         //Aqui chamaremos na api, os produtos
         return(
-        <div key={e.id} className="produto">
+          
+        <Link key={e.id_produto} className="produto" to={`/editar/${e.id_produto}`}>
           <BiPackage size="73px" />
-          <h1>{e.nome}</h1>
-        </div>
+          <h1>{e.produto}</h1>
+            <FaTrash style={{cursor: 'pointer'}} onClick={()=>DeletarProduto(e.id_produto)} size="20px" />
+        </Link>
         )
         }) }
       </div>
@@ -78,10 +101,10 @@ export default function AddProdutos() {
               <div className="dirCard">
                 <h4>Produtos mais vendidos</h4>
                 
-                {data.map(e=>{
+                {/*data.map(e=>{
                   //Aqui chamaremos da api de estatisticas do produto
                   return(<span>{e.nome}</span>)
-                })}
+                })*/}
               </div>
 
               <div className="esqCard">